rtic_common::wait_queue

Struct DoublyLinkedList

pub struct DoublyLinkedList<T> { /* private fields */ }

An atomic, doubly linked, FIFO list for a wait queue.

Atomicity is guaranteed by short critical_sections, so this list is not lock free, but it will not deadlock.

Implementations

impl<T> DoublyLinkedList<T>

pub const fn new() -> Self

Create a new linked list.

impl<T: Clone> DoublyLinkedList<T>

pub fn pop(&self) -> Option<T>

Pop the first element in the queue.

pub unsafe fn push(&self, link: Pin<&Link<T>>)

Put an element at the back of the queue.

Safety

The link must live until it is removed from the queue.

pub fn is_empty(&self) -> bool

Check if the queue is empty.

impl DoublyLinkedList<Waker>

pub async fn wait_until<T, F: FnMut() -> Option<T>>(&self, f: F) -> T

Wait until f returns Some.